defaulting modal height if default index is selected
my tests who the issue is indeed with modal-dropdown itself, whenever a default index is selected and modal has a custom height, the scroll is messed up....

as a quick fix, whenever there is a pre selected index, I set the height to default value, this hack fixes the issue for use for now.

The real fix would be to update the modal dropdown itself.